splitting flexindex into multiple files
