+++ pyterrier_dr/__init__.py
@@ -1,6 +1,6 @@
 from .util import SimFn
 from .indexes import DocnoFile, NilIndex, NumpyIndex, RankedLists, FaissFlat, FaissHnsw, MemIndex, TorchIndex
-from .flex_index import FlexIndex
+from .flex import FlexIndex
 from .biencoder import BiEncoder, BiQueryEncoder, BiDocEncoder, BiScorer
 from .hgf_models import HgfBiEncoder, TasB, RetroMAE
 from .sbert_models import SBertBiEncoder, Ance, Query2Query
