+++ pyterrier_dr/flex/__init__.py
@@ -1,0 +1,3 @@
+from .core import FlexIndex
+from .corpus_graph import *
+from .faiss_retr import *
